+++ src/AcadMcp.Plugin/PluginEntryPoint.cs
@@ -69,6 +69,7 @@ public sealed class PluginEntryPoint : AcadRuntime.IExtensionApplication
             DataLinkPluginTools.Register(_toolHost);
             ViewsPluginTools.Register(_toolHost);
             SelectionExtPluginTools.Register(_toolHost);
+            GeoPluginTools.Register(_toolHost);
             BooleanOpsPluginTools.Register(_toolHost);
             ModifyPluginTools.Register(_toolHost);
             SelectionPluginTools.Register(_toolHost);